# backend/swagger.json
{"openapi":"3.0.0","paths":{"/api/auth/sign-up":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDto"}}}},"responses":{"200":{"description":"Successfully registered","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/auth/login":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDto"}}}},"responses":{"200":{"description":"Get user details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginPayloadDto"}}}}}}},"/api/auth/profile":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":"Get user details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/auctions":{"post":{"operationId":"AuctionController_createAuction","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAuctionDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"AuctionController_getAuctions","parameters":[],"responses":{"200":{"description":"Get actions","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AuctionDto"}}}}}}}},"/api/auctions/{id}":{"get":{"operationId":"AuctionController_getAuction","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"AuctionController_updateAuction","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAuctionDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"AuctionController_deleteAuction","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}},"/api/auctions/{id}/bid":{"post":{"operationId":"BidController_placeBid","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}}},"get":{"operationId":"BidController_getBids","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"API","description":"","version":"0.0.1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"UserLoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"UserDto":{"type":"object","properties":{"id":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"email":{"type":"string"}},"required":["id","createdAt","updatedAt","email"]},"LoginPayloadDto":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserDto"},"accessToken":{"type":"string"}},"required":["user","accessToken"]},"CreateAuctionDto":{"type":"object","properties":{}},"AuctionDto":{"type":"object","properties":{"id":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"title":{"type":"string"},"description":{"type":"string"},"currentPrice":{"type":"number"},"endTime":{"format":"date-time","type":"string"}},"required":["id","createdAt","updatedAt","title","description","currentPrice","endTime"]},"UpdateAuctionDto":{"type":"object","properties":{}}}}}